[Story #32383613] - Charts for Manager View
- Use floatval() instead of intval() as there are potentials for float values.

+++ sugarcrm/modules/Forecasts/api/ForecastsChartApi.php
@@ -186,7 +186,7 @@ class ForecastsChartApi extends ChartApi
             }
 
             // set the variables
-            $dataArray['values'][$key]['goalmarkervalue'] = array(intval($quota['amount']), intval($likely));
+            $dataArray['values'][$key]['goalmarkervalue'] = array(floatval($quota['amount']), floatval($likely));
             $dataArray['values'][$key]['goalmarkervaluelabel'] = array($quota['formatted_amount'], $likely_label);
         }
